Moved the java serialization files to core-java package
