+++ core-java/src/test/java/com/baeldung/javaSerialization/externalizable/ExternalizationUnitTest.java
@@ -1,4 +1,4 @@
-package com.java.test.externalizable;
+package com.baeldung.javaSerialization.externalizable;
 
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -6,8 +6,8 @@ import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 
-import com.java.test.externalizable.Employee;
-import com.java.test.externalizable.ExternalizationUnitTest;
+import com.baeldung.javaSerialization.externalizable.Employee;
+import com.baeldung.javaSerialization.externalizable.ExternalizationUnitTest;
 
 import org.junit.Test;
 import org.junit.*;
@@ -18,7 +18,7 @@ import org.junit.*;
  */
 public class ExternalizationUnitTest {
 
-    private String filePath = "employee.ser";
+    private String filePath = "com/baeldung/javaSerialization/externalizable/employee.ser";
 
     public Employee serialize() throws IOException {
         Employee emp = new Employee();
@@ -44,7 +44,7 @@ public class ExternalizationUnitTest {
     }
 
     @Test
-    public void externalizableTest() throws Exception {
+    public void givenEmployee_whenSerializedOrDeserialized_thenOk() throws Exception {
         ExternalizationUnitTest externalizationExample = new ExternalizationUnitTest();
         Employee empAfterSerialization = externalizationExample.serialize();
         Assert.assertEquals(empAfterSerialization.id, 73737);
