--- java-serialization/README.md
@@ -1,2 +1,0 @@
-### Relevant Articles:
-- [Introduction to Java Serialization](http://inprogress.baeldung.com/introduction-to-java-serialization/)
